+++ src/indi/zhaiaixn/blog/dao/impl/BlogDaoImpl.java
@@ -58,7 +58,7 @@ public class BlogDaoImpl implements BlogDao{
     }
 
 
-    public boolean delBlog(long id){
+    public boolean delBlog(int id){
 
         boolean flag = false;
         connection = null;
@@ -93,7 +93,7 @@ public class BlogDaoImpl implements BlogDao{
             ResultSet rs = pstmt.executeQuery();
             while (rs.next()){
                 blog = new Blog();
-                blog.setId(rs.getLong("id"));
+                blog.setId(rs.getInt("id"));
                 blog.setTitle(rs.getString("title"));
                 blog.setAbstracts(rs.getString("abstracts"));
                 blog.setContent(rs.getString("content"));
@@ -121,7 +121,7 @@ public class BlogDaoImpl implements BlogDao{
             ResultSet rs = pstmt.executeQuery();
             while (rs.next()){
                 blog = new Blog();
-                blog.setId(rs.getLong("id"));
+                blog.setId(rs.getInt("id"));
                 blog.setTitle(rs.getString("title"));
                 blog.setAbstracts(rs.getString("abstracts"));
                 blog.setContent(rs.getString("content"));
@@ -149,7 +149,7 @@ public class BlogDaoImpl implements BlogDao{
             ResultSet rs = pstmt.executeQuery();
             while (rs.next()){
                 blog = new Blog();
-                blog.setId(rs.getLong("id"));
+                blog.setId(rs.getInt("id"));
                 blog.setTitle(rs.getString("title"));
                 blog.setAbstracts(rs.getString("abstracts"));
                 blog.setContent(rs.getString("content"));
@@ -181,4 +181,56 @@ public class BlogDaoImpl implements BlogDao{
         }
         return categoryList;
     }
+
+//    public List<Blog> listAllPageing(String pageIndex,String pageSize,int totalPage){
+//
+//        if(pageIndex==null||"".equals(pageIndex.trim())){//trim()去掉字符串首尾的空格
+//            pageIndex="1";
+//        }
+//
+//        if(pageSize==null||"".equals(pageSize.trim())){
+//            pageSize="5";
+//        }
+//        int start =(Integer.parseInt(pageIndex)-1)*Integer.parseInt(pageSize);
+//        //数据库中所有数据的数量
+//        int count=0;
+//        //总的页数
+//
+//
+//        List<Blog> blogList = new ArrayList<Blog>();
+//        PreparedStatement pstmt = null;
+//        Blog blog = null;
+//
+//        try{
+//            this.connection = connPool.getConnection();
+//            pstmt = connection.prepareStatement(BlogSql.countBlog);
+//            ResultSet rsCount = pstmt.executeQuery();
+//            while (rsCount.next()){
+//              count = rsCount.getInt("count");
+//            }
+//
+//            totalPage = count/Integer.parseInt(pageSize);//总页数=总记录数/每页的记录数
+//
+//            totalPage = count%Integer.parseInt(pageSize)>0?(totalPage+1):totalPage;//若有余数，则总页数加一；否则不变
+//
+//
+//            pstmt = connection.prepareStatement(BlogSql.listBlogPageing);
+//            pstmt.setInt(1,start);
+//            pstmt.setInt(2,Integer.parseInt(pageSize));
+//            ResultSet rs = pstmt.executeQuery();
+//            while (rs.next()){
+//                blog = new Blog();
+//                blog.setId(rs.getInt("id"));
+//                blog.setTitle(rs.getString("title"));
+//                blog.setAbstracts(rs.getString("abstracts"));
+//                blog.setContent(rs.getString("content"));
+//                blog.setCategory(rs.getString("category"));
+//                blogList.add(blog);
+//            }
+//        } catch (SQLException e) {
+//            e.printStackTrace();
+//        }
+//
+//        return blogList;
+//    }
 }
